+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.5.14 - Unreleased
 
+### Added
+* Add `argsort` to ArrayAPI.
+
 
 ## Version 0.5.13 - Released 2021-01-08